Prevent circular imports
This moves code around so we no longer have a circular import.
