+++ luigi/tasks/quickgo/quickgo.py
@@ -1,0 +1,29 @@
+# -*- coding: utf-8 -*-
+
+"""
+Copyright [2009-2017] EMBL-European Bioinformatics Institute
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+
+import luigi
+
+from .quickgo_data import QuickGoData
+from .pgload_quickgo import PgLoadQuickGo
+
+
+class QuickGo(luigi.WrapperTask):  # pylint: disable=R0904
+    """
+    Parse the QuickGO data and import into the database.
+    """
+
+    def requires(self):
+        yield QuickGoData()
+        yield PgLoadQuickGo()
